refactor: extract all style related code into useChipStyles
