+++ src/components/_global/BalChip/composables/useChipStyles.ts
@@ -1,6 +1,6 @@
 import { computed } from 'vue';
 
-export default function useChipClasses(props) {
+export function useChipClasses(props) {
   const sizeClasses = () => {
     switch (props.size) {
       case 'sm':
@@ -41,4 +41,28 @@ export default function useChipClasses(props) {
       [shapeClasses()]: true
     };
   });
+}
+
+export function useCloseIconClasses(props) {
+  const isGradient = props.color === 'gradient';
+  const colorClass = isGradient ? 'text-white' : `text-${props.color}-500`;
+
+  const classes = computed(() => {
+    return {
+      [colorClass]: true
+    };
+  });
+
+  const iconSize = computed(() => {
+    switch (props.size) {
+      case 'sm':
+        return 'xxs';
+      case 'lg':
+        return 'sm';
+      default:
+        return 'xs';
+    }
+  });
+
+  return { classes, iconSize };
 }
